requirements.txt
Flask
google-analytics-data
google-auth
google-auth-oauthlib
google-auth-httplib2
grpcio
